Fresh Start
I have decided to completely wipe the repository because I had so many issues that were made apparent that the technical debt rose very suddenly as I realised that the architecture I designed was not good enough for the project I am writing

This is most notable in the multithreaded aspect of the engine. glfw is as far as I am aware, not multithreaded. I didn't know this because I remember when working on Windows with the glfw binaries, there was a separate glfw3_mt.lib version.
 This I think was actually for MSVC's multithreaded runtime library, rather than glfw itself being multithreaded

 I am also not happy about the way the engine is initialised. Currently, the engine is initialised when a client application class is constructed. This is bad because when the client app is created the engine can be initialised multiple times. This was circumvented by adding an m_isInitialised variable, but that is just a stupid way to fix a stupid problem.